refactor: simplify and comment for wilson charts
