--- premade-charts/src/lib/Hover.svelte
@@ -1,19 +1,0 @@
-<script lang="ts">
-    import { createEventDispatcher } from "svelte";
-    const dispatch = createEventDispatcher();
-</script>
-
-<div
-    id="hover-container"
-    on:mouseenter={() => dispatch("hover", true)}
-    on:mouseleave={() => dispatch("hover", false)}
->
-    <slot />
-</div>
-
-<style>
-    #hover-container {
-        display: inline-block;
-        outline: 1px solid black;
-    }
-</style>
